Add PoolState to the LunInfo table
Add a new column that tracks the PoolState (from the  SFAStoragePool
class) for all LUNs.  The value is an integer - see the SFA OS API Guide
for the meaning of the integer.

Also removed the code that was reading the SFADiskDriveStatistics
objects at the medium rate.  We were never actually doing anything with
that data.