Go back to using in-memory tables
For reasons that are currently unclear, the switch to using INNODB
tables has resulted in no data actually making into the database.  (It's
possible it's something to do with transactions not commiting, but
that's just a guess.)  Until we know what's going on, I'm going back to
using HEAP tables.