Deprecate the 'database' section of the config file
Replace the 'database' section of the config file with 'SqlDb'.  Use
of the 'database' section is deprecated: the code will still read
the entries from that section, but will log a warning.  Also, if both
'SqlDb' and 'database' exist in the same config file, 'database'
will be ignored (though the code will log a warning about this, too).

This is in preparation for adding a section for a time-series database.