Fix handling of [database] section of config file
The [database] section is deprecated and replaced by the [SqlDb]
section, yet DDNTool.py still expected to find [database].  Fixed that.
Also some other minor cleanups: turn down the logging verbosity for
the request package (which is used by the influxdb package) and print
out the actual stack trace when the main process detects an unhandled
exception in any of the child processes.