Change InfluxDB schema
Change the schema for the histogram data in SFAInfluxDB.  Instead of one
measurement with a field for each bucket, there's now a "bucket" tag and
we'll have multiple measurements each with a single value.  This lets us
used the GROUP BY clause more effectively.  (Still don't know how to plot
a histogram in grafana, though...)