Fix minor race condition with update time
When the user hit's crtl-C to exit, the main process sets the update
time to 0.  When the sub-processes see this 0, they exit.  However,
the main process could set the time to 0 while a subprocess is in the
middle of its main loop.  In such a case, 0 could get written to the
database.

Solved the problem by having the subprocesses copy the time value from
the shared-mem location to a local variable at the top of their main
loops.  Thereafter, they only refer to the local time.