Reduce calls to InfluxDB server
Modify SFAClient and SFAInfluxDb classes to queue up data and only dump
it to the db server once per call to the *_tsdb_tasks() functions. This
reduces the number of HTTP connections to the DB server by about a
factor or 20 (though with a corresponding 20X increase in the amount of
data sent per connection).  This reduces the load on the db server (and
probably also on the server running DDNTool).