Moving MANIFEST.in out of the way
It looks like we no longer need MANIFEST.in.  (We were using it as a work-
around for what appeared to be a bug in the RHEL6 version of distutils.  With
the move to setuptools, we don't have that bug any more.)  Just renaming the
file for now.  Once I'm sure I don't need it, I'll delete it completely.