Support for DDN firmware version 3 and higher
Updates to support the v3.x firmware from DDN.  The main difference between v2
and v3 is that the column labels for the latency table data have changed.  This
requires changes to the _check_labels() and to the SQL DB initialization code.
Also, in order know what DB tables to create, we need to create an instance of
SFAClient and get the firmware version.  This required some minor changes to
SFAClient.