Reduce logging verbosity
DDNTool.py had code that was supposed to stop excessive log messages on the
ddn.sfa.core.* loggers, but it wasn't working properly.  Fixed that. Also
removed the references to the 'pywebm' and 'root' loggers because they
weren't logging anything anyway.  Upgraded a couple of messages in SFAClient
from debug to info so that we can still tell at a glance that the background
processes are still running.

(Changes are in response to a pull request from user treydock.)

Also bumped the version to 2.6.2